+++ src/app/components/svgs/ZZLogo/story.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { IATILogoColor, IATILogoWhite } from '.';
+import { ZZLogo } from '.';
 import Providers from 'app/Providers';
 
 storiesOf('SVG,s', module).add('IATI Logos', () => (
   <Providers>
-    <IATILogoColor />
-    <IATILogoWhite />
+    <ZZLogo />
   </Providers>
 ));
+
