+++ src/app/modules/QueryBuilder/layout.tsx
@@ -1,22 +1,21 @@
 /* core */
-import React from 'react';
 /* third-party */
 import Container from '@material-ui/core/Container';
-import Box from '@material-ui/core/Box';
-/* common */
-import { FragmentDivider } from 'app/modules/QueryBuilder/fragments/common';
-/* fragments */
-import {
-  DownloadFragment,
-  FilterFragment,
-  IntroFragment,
-  OutputFragment,
-  ResultFragment,
-  WhoFragment,
-} from 'app/modules/QueryBuilder/fragments';
-import { ModuleStore } from 'app/modules/QueryBuilder/state/store';
-import { PageContainer } from 'app/modules/common/PageContainer';
 import { Header } from 'app/components/surfaces/Header';
+import { PageContainer } from 'app/modules/common/PageContainer';
+/* fragments */
+import { IntroFragment } from 'app/modules/QueryBuilder/fragments';
+import { ModuleStore } from 'app/modules/QueryBuilder/state/store';
+import React from 'react';
+import { QueryBuilderStepper } from './common/stepper/layout';
+import { Grid } from '@material-ui/core';
+import { Skeletor } from 'app/components/utils/Skeletor';
+import { OrganisationFragment } from 'app/modules/QueryBuilder/steps/organisation';
+import { DebugBox } from 'app/utils/layout';
+import { Route, Switch } from 'react-router';
+import { AdditionalFiltersLayout } from 'app/modules/QueryBuilder/steps/additional-filters/layout';
+import { OutputFormatLayout } from 'app/modules/QueryBuilder/steps/output-format/layout';
+import { ResultsLayout } from 'app/modules/QueryBuilder/steps/results/layout';
 /* config & mock */
 
 export const QueryBuilderLayout = () => {
@@ -25,30 +24,69 @@ export const QueryBuilderLayout = () => {
   return (
     <PageContainer>
       <Header>
-        <IntroFragment/>
-          {/*<FragmentDivider />*/}
+        <IntroFragment />
+        {/*<FragmentDivider />*/}
       </Header>
 
-      <Container maxWidth="lg">
-        {/* WHO FRAGMENT */}
-        <WhoFragment store={store} />
-        <FragmentDivider />
 
-        {/* FILTERS FRAGMENT */}
-        <FilterFragment />
-        <FragmentDivider />
+        {/* step navigator */}
+        <Container maxWidth="lg">
+          {/*Grid container spacing={4}>
+            <Grid item xs={12}>
+              <Skeletor height="200px" width="100%" />
+            </Grid>
+          </Grid>*/
 
-        {/* OUTPUT FORMAT FRAGMENT */}
-        <OutputFragment />
-        <FragmentDivider />
+          /* step content */
 
-        {/* RESULT FRAGMENT */}
-        {/* the ResultFragment contains FragmentDivider */}
-        <ResultFragment />
+          /*<Switch>
+            <Route
+              exact
+              path="/querybuilder/organisation"
+              render={() => <OrganisationFragment />}
+            />
 
-        {/* DOWNLOAD FRAGMENT */}
-        <DownloadFragment store={store} />
-      </Container>
+            <Route
+              exact
+              path="/querybuilder/filters"
+              render={() => <AdditionalFiltersLayout />}
+            />
+
+            <Route
+              exact
+              path="/querybuilder/output"
+              render={() => <OutputFormatLayout />}
+            />
+
+            <Route
+              exact
+              path="/querybuilder/results"
+              render={() => <ResultsLayout />}
+            />
+          </Switch>*/}
+
+          <QueryBuilderStepper />
+
+          {/* WHO FRAGMENT */}
+          {/*<WhoFragment store={store} />*/}
+          {/*<FragmentDivider />*/}
+
+          {/* FILTERS FRAGMENT */}
+          {/*<FilterFragment />*/}
+          {/*<FragmentDivider />*/}
+
+          {/* OUTPUT FORMAT FRAGMENT */}
+          {/*<OutputFragment />*/}
+          {/*<FragmentDivider />*/}
+
+          {/* RESULT FRAGMENT */}
+          {/* the ResultFragment contains FragmentDivider */}
+          {/*<ResultFragment />*/}
+
+          {/* DOWNLOAD FRAGMENT */}
+          {/*<DownloadFragment store={store} />*/}
+        </Container>
+
     </PageContainer>
   );
 };
