+++ src/app/modules/QueryBuilder/steps/additional-filters/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { FilterFragment } from 'app/modules/QueryBuilder/fragments';
+
+export function AdditionalFiltersSubModule() {
+  return (
+    <>
+      <FilterFragment />
+    </>
+  );
+}
