+++ src/app/modules/QueryBuilder/steps/output-format/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { OutputFragment } from 'app/modules/QueryBuilder/fragments';
+
+export function OutputFormatSubModule() {
+  return (
+    <>
+      <OutputFragment />
+    </>
+  );
+}
