+++ src/app/modules/QueryBuilder/steps/results/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { DownloadFragment } from 'app/modules/QueryBuilder/fragments';
+
+export function ResultsSubModule() {
+  return (
+    <>
+      <DownloadFragment />
+    </>
+  );
+}
